+++ include/hybridmap/HybridSynthesisMapper.hpp
@@ -20,9 +20,9 @@
 #include "hybridmap/Mapping.hpp"
 #include "hybridmap/NeutralAtomDefinitions.hpp"
 #include "hybridmap/NeutralAtomScheduler.hpp"
-#include "hybridmap/OpenQASMSerializer.hpp"
 #include "ir/Definitions.hpp"
 #include "ir/QuantumComputation.hpp"
+#include "na/ir/OpenQASMSerializer.hpp"
 
 #include <cstdint>
 #include <fstream>