+++ src/sc/Mapper.cpp
@@ -55,6 +55,7 @@ void Mapper::initResults() {
   results.output.name = qc.getName() + "_mapped";
   results.output.qubits = architecture->getNqubits();
   results.output.gates = std::numeric_limits<std::size_t>::max();
+  qcMapped.setName(qc.getName());
   qcMapped.addQubitRegister(architecture->getNqubits());
 }